requirements.txt
newthon
pymongo
dnspython
python-dotenv
speedtest-cli
psutil
requests 
cryptg
seedrcc
pillow 
